+++ src/components/input-time-picker/input-time-picker.tsx
@@ -93,7 +93,7 @@ export class InputTimePicker
   /**
    * The ID of the form that will be associated with the component.
    *
-   * When not set, the component will be associated with its ancestor `<form>` element, if any.
+   * When not set, the component will be associated with its ancestor form element, if any.
    */
   @Prop({ reflect: true })
   form: string;
@@ -199,7 +199,8 @@ export class InputTimePicker
   @State() effectiveLocale = "";
 
   @Watch("effectiveLocale")
-  effectiveLocaleWatcher(): void {
+  @Watch("step")
+  valueRelatedPropChange(): void {
     this.setInputValue(
       localizeTimeString({
         value: this.value,